some rpcs and toasts

+++ src/lib/projects.remote.ts
@@ -1,15 +1,23 @@
-import { command, form, query } from '$app/server';
+import { command, form, getRequestEvent, query } from '$app/server';
 import { db } from '$lib/server/db';
+import { validator } from 'svelte-checkmate';
 import { project, projectStatusEnum } from '$lib/server/db/schema';
-import { getUser } from '$lib/server/server-utils';
-import { error } from '@sveltejs/kit';
+import { error, redirect } from '@sveltejs/kit';
 import { and, eq } from 'drizzle-orm';
 import { z } from 'zod';
 import { zEditProject, zNewProject } from './zod';
 // import { STATUS_VALUES } from './utils';
 
+export const getUser = query(async () => {
+	const { locals } = getRequestEvent();
+
+	if (!locals.user) redirect(302, '/home');
+
+	return locals.user;
+});
+
 export const getProjects = query(async () => {
-	const user = getUser();
+	const user = await getUser();
 	const projects = db.query.project.findMany({
 		where: {
 			ownerId: user.id
@@ -24,7 +32,7 @@ export const getProjects = query(async () => {
 });
 
 export const getProject = query(z.number(), async (id: number) => {
-	const user = getUser();
+	const user = await getUser();
 	const project = await db.query.project.findFirst({
 		where: {
 			ownerId: user.id,
@@ -37,42 +45,29 @@ export const getProject = query(z.number(), async (id: number) => {
 	return project;
 });
 
-export const statusVals = query(async () => {
-	return projectStatusEnum.enumValues;
-});
-
 export const newProject = form(async (formData) => {
-	const user = getUser();
-	const values = {
-		title: formData.get('title')
-	};
-	const result = zNewProject.safeParse(values);
+	const user = await getUser();
 
-	if (!result.success) return { errors: result.error.flatten().fieldErrors };
+	const result = await validator({ schema: zNewProject, formData });
 
-	const { title } = result.data;
+	if (!result.success) return error(400, result.errors.title);
 
 	await db.insert(project).values({
-		title: title,
+		title: result.data.title,
 		ownerId: user.id
 	});
 
-	return { success: true };
+	return { success: true, title: result.data.title };
 });
 
 export const editProject = form(async (formData) => {
-	const user = getUser();
+	const user = await getUser();
 
-	// check values
-	const values = {
-		projectId: Number(formData.get('id')),
-		title: formData.get('title')
-	};
-	const result = zEditProject.safeParse(values);
+	const result = await validator({ schema: zEditProject, formData });
 
-	if (!result.success) return { errors: result.error.flatten().fieldErrors };
+	if (!result.success) return error(400, result.errors.newTitle);
 
-	const { title, projectId } = result.data;
+	const { newTitle, projectId } = result.data;
 
 	// query db to check if record exists
 	const record = await db.query.project.findFirst({
@@ -82,15 +77,15 @@ export const editProject = form(async (formData) => {
 		}
 	});
 
-	if (!record) return { success: false, error: 'Project not found or permission denied' };
+	if (!record) return error(404, 'Project not found or permission denied');
 
 	await db
 		.update(project)
 		.set({
-			title
+			title: newTitle
 		})
 		.where(eq(project.id, projectId));
-	return { success: true };
+	return { success: true, title: newTitle };
 });
 
 export const updateDate = command(
@@ -108,7 +103,7 @@ export const updateDate = command(
 );
 
 export const clearDate = command(z.number(), async (projectId) => {
-	const user = getUser();
+	const user = await getUser();
 
 	await db
 		.update(project)
@@ -121,7 +116,7 @@ export const clearDate = command(z.number(), async (projectId) => {
 export const drop = command(
 	z.object({ id: z.number(), target: z.enum(projectStatusEnum.enumValues) }),
 	async ({ id, target }) => {
-		const user = getUser();
+		const user = await getUser();
 
 		const qProject = await db.query.project.findFirst({
 			where: {
@@ -130,7 +125,7 @@ export const drop = command(
 			}
 		});
 
-		if (!qProject) return error(404, 'Project not found');
+		if (!qProject) return error(404, 'Project not found or permission denied');
 		await db
 			.update(project)
 			.set({
@@ -141,7 +136,7 @@ export const drop = command(
 );
 
 export const deleteProject = command(z.number(), async (id) => {
-	const user = getUser();
+	const user = await getUser();
 
 	const qProject = await db.query.project.findFirst({
 		where: {
@@ -150,7 +145,7 @@ export const deleteProject = command(z.number(), async (id) => {
 		}
 	});
 
-	if (!qProject) return error(404, 'Project not found');
+	if (!qProject) return error(404, 'Project not found or permission denied');
 
 	await db.delete(project).where(eq(project.id, qProject.id));
 });
